tests/filters/image/watermark/requirements.txt
# Базовые зависимости DEPP
boto3==1.35.6
botocore==1.35.6
numpy==1.26.4

# зависимости фильтра
onnx==1.16.2
onnxruntime==1.18
opencv-python-headless==4.8.1.78
pandas==2.2.2
pillow==10.4.0
pillow_heif==0.18.0
PyYAML==6.0.2
redis==5.1.1
# Зависимости utils
requests==2.32.3
scenedetect==0.6.4
